fix(xla): make Molmo sparse merge atomic
Preflight every destination sum so non-finite overflow cannot partially mutate a reusable batch slot, and reject processor mask values outside the qualified [-1,1] coverage-and-sentinel domain.

Keep raw patch-mask rows distinct from pooled feature rows and split tests out of the production module to preserve file-size limits.

Refs #870
